renames CodePreview back to CodeBlock
